requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pytorch-lightning>=2.2.0
timm>=0.9.12
hydra-core>=1.3.2
omegaconf>=2.3.0
tensorboard>=2.15.0
wandb>=0.16.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
albumentations>=1.4.0
opencv-python>=4.9.0
rasterio>=1.3.9
geopandas>=0.14.0
shapely>=2.0.2
matplotlib>=3.8.0
tqdm>=4.66.0
pyproj